fix formating

+++ smart-contract/lottery/src/Lottery.sol
@@ -18,7 +18,7 @@ contract Lottery {
     address[] private s_participantes;
     address private s_lastRoundWinner;
     address private s_owner;
-    uint private s_lastRoundStartedAt;
+    uint256 private s_lastRoundStartedAt;
     Status private s_lotteryStatus;
 
     uint256 private immutable i_entryFee;
@@ -38,7 +38,7 @@ contract Lottery {
         return s_lotteryStatus;
     }
 
-    function getLastRoundStarted() external view returns (uint) {
+    function getLastRoundStarted() external view returns (uint256) {
         return s_lastRoundStartedAt;
     }
 
@@ -78,15 +78,14 @@ contract Lottery {
         s_lotteryStatus = Status.CALCULATING;
 
         //TODO: Use VRF of chainlink
-        uint winnerIndex = s_participantes.length / 2;
+        uint256 winnerIndex = s_participantes.length / 2;
 
         s_lastRoundWinner = s_participantes[winnerIndex];
 
-        uint256 platformComision = (address(this).balance *
-            PLATFORM_COMMISION_IN_PERCENTAGE) / 100;
+        uint256 platformComision = (address(this).balance * PLATFORM_COMMISION_IN_PERCENTAGE) / 100;
         uint256 winningPrize = address(this).balance - platformComision;
 
-        (bool success, ) = s_lastRoundWinner.call{value: winningPrize}("");
+        (bool success,) = s_lastRoundWinner.call{value: winningPrize}("");
 
         if (!success) {
             revert Lottery_PaymentFailed();
